tambah sequence dan merubah struktur class dan file yang berbau mrp
