--- data/mrp_serial_number_sequence.xml
@@ -1,10 +1,0 @@
-<odoo>
-    <data noupdate="1">
-    	<record forcecreate="True" id="mrp_serial_number_sequence" model="ir.sequence">
-		    <field name="name">Mrp Serial Number Sequence</field>
-		    <field name="code">mrp.serial.number</field>
-		    <field name="prefix">ICE-%(y)s%(month)s-</field>
-		    <field name="padding">4</field>
-		</record>
-    </data>
-</odoo>
